// src/utils/enums/ticket.enum.ts
export enum TicketTypeEnum {
  BUG = "Bug",
  TASK = "Task",
  MR = "MR Review",
  STORY = "Story",
  SPIKE_STORY = "Spike Story",
  SPIKE_TASK = "Spike Task",
}

export enum MRType {
  MERGED = "Merged",
  IN_PROGRESS = "In Progress",
  IN_REVIEW = "In Review",
  Dash = "-",
}

export enum StatusType {
  TODO = "To Do",
  IN_PROGRESS = "In Progress",
  IN_REVIEW = "In Review",
  RESOLVED = "Resolved",
  READY_FOR_VERIFICATION = "Ready For Verification",
  READY_FOR_ACCEPTANCE = "Ready For Acceptance",
  BLOCKED = "Blocked",
}